+++ Sources/ZcashLightClientKit/Tor/TorClient.swift
@@ -12,7 +12,7 @@ public class TorClient {
     private let runtime: OpaquePointer
     public var cachedFiatCurrencyResult: FiatCurrencyResult?
 
-    init(torDir: URL) async throws {
+    init(torDir: URL) throws {
         // Ensure that the directory exists.
         let fileManager = FileManager()
         if !fileManager.fileExists(atPath: torDir.path) {
@@ -41,20 +41,19 @@ public class TorClient {
         zcashlc_free_tor_runtime(runtime)
     }
 
-    public func isolatedClient() async throws -> TorClient {
+    public func isolatedClient() throws -> TorClient {
         let isolatedPtr = zcashlc_tor_isolated_client(runtime)
 
         guard let isolatedPtr else {
             throw ZcashError.rustTorIsolatedClient(
-                lastErrorMessage(
-                    fallback:
-                        "`TorClient.isolatedClient` failed with unknown error"))
+                lastErrorMessage(fallback: "`TorClient.isolatedClient` failed with unknown error")
+            )
         }
 
         return TorClient(runtimePtr: isolatedPtr)
     }
 
-    public func getExchangeRateUSD() async throws -> FiatCurrencyResult {
+    public func getExchangeRateUSD() throws -> FiatCurrencyResult {
         let rate = zcashlc_get_exchange_rate_usd(runtime)
 
         if rate.is_sign_negative {
@@ -64,8 +63,10 @@ public class TorClient {
         let newValue = FiatCurrencyResult(
             date: Date(),
             rate: NSDecimalNumber(
-                mantissa: rate.mantissa, exponent: rate.exponent,
-                isNegative: rate.is_sign_negative),
+                mantissa: rate.mantissa,
+                exponent: rate.exponent,
+                isNegative: rate.is_sign_negative
+            ),
             state: .success
         )
 
@@ -74,23 +75,19 @@ public class TorClient {
         return newValue
     }
 
-    public func connectToLightwalletd(endpoint: String) async throws
-        -> TorLwdConn
-    {
+    public func connectToLightwalletd(endpoint: String) throws -> TorLwdConn {
         guard !endpoint.containsCStringNullBytesBeforeStringEnding() else {
-            throw ZcashError.rustTorConnectToLightwalletd(
-                "endpoint string contains null bytes")
+            throw ZcashError.rustTorConnectToLightwalletd("endpoint string contains null bytes")
         }
 
         let lwdConnPtr = zcashlc_tor_connect_to_lightwalletd(
-            runtime, [CChar](endpoint.utf8CString))
+            runtime, [CChar](endpoint.utf8CString)
+        )
 
         guard let lwdConnPtr else {
             throw ZcashError.rustTorConnectToLightwalletd(
-                lastErrorMessage(
-                    fallback:
-                        "`TorClient.connectToLightwalletd` failed with unknown error"
-                ))
+                lastErrorMessage(fallback: "`TorClient.connectToLightwalletd` failed with unknown error")
+            )
         }
 
         return TorLwdConn(connPtr: lwdConnPtr)
@@ -100,6 +97,7 @@ public class TorClient {
 public class TorLwdConn {
     private let conn: OpaquePointer
 
+    // swiftlint:disable:next strict_fileprivate
     fileprivate init(connPtr: OpaquePointer) {
         conn = connPtr
     }
@@ -111,9 +109,7 @@ public class TorLwdConn {
     /// Submits a raw transaction over lightwalletd.
     /// - Parameter spendTransaction: data representing the transaction to be sent
     /// - Throws: `serviceSubmitFailed` when GRPC call fails.
-    func submit(spendTransaction: Data) async throws
-        -> LightWalletServiceResponse
-    {
+    func submit(spendTransaction: Data) throws -> LightWalletServiceResponse {
         let success = zcashlc_tor_lwd_conn_submit_transaction(
             conn,
             spendTransaction.bytes,
@@ -121,18 +117,21 @@ public class TorLwdConn {
         )
 
         var response = SendResponse()
+        
         if !success {
-            let err = lastErrorMessage(
-                fallback: "`TorLwdConn.submit` failed with unknown error")
-            if err.hasPrefix("Failed to submit transaction (")
-                && err.contains(")")
-            {
-                let startOfCode = err.firstIndex(of: "(")!
-                let endOfCode = err.firstIndex(of: ")")!
-                let errorCode = Int32(
-                    err[err.index(startOfCode, offsetBy: 1)..<endOfCode])!
-                let errorMessage = String(
-                    err[err.index(endOfCode, offsetBy: 3)...])
+            let err = lastErrorMessage(fallback: "`TorLwdConn.submit` failed with unknown error")
+            
+            if err.hasPrefix("Failed to submit transaction (") && err.contains(")") {
+                guard let startOfCode = err.firstIndex(of: "(") else {
+                    throw ZcashError.rustTorLwdSubmit(err)
+                }
+                guard let endOfCode = err.firstIndex(of: ")") else {
+                    throw ZcashError.rustTorLwdSubmit(err)
+                }
+                guard let errorCode = Int32(err[err.index(startOfCode, offsetBy: 1)..<endOfCode]) else {
+                    throw ZcashError.rustTorLwdSubmit(err)
+                }
+                let errorMessage = String(err[err.index(endOfCode, offsetBy: 3)...])
 
                 response.errorCode = errorCode
                 response.errorMessage = errorMessage
@@ -140,6 +139,7 @@ public class TorLwdConn {
                 throw ZcashError.rustTorLwdSubmit(err)
             }
         }
+        
         return response
     }
 
@@ -148,9 +148,7 @@ public class TorLwdConn {
     /// - Throws: LightWalletServiceError
     /// - Returns: LightWalletServiceResponse
     /// - Throws: `serviceFetchTransactionFailed` when GRPC call fails.
-    func fetchTransaction(txId: Data) async throws -> (
-        tx: ZcashTransaction.Fetched?, status: TransactionStatus
-    ) {
+    func fetchTransaction(txId: Data) throws -> (tx: ZcashTransaction.Fetched?, status: TransactionStatus) {
         guard txId.count == 32 else {
             throw ZcashError.rustGetMemoInvalidTxIdLength
         }
@@ -161,10 +159,7 @@ public class TorLwdConn {
 
         guard let txPtr else {
             throw ZcashError.rustTorLwdFetchTransaction(
-                lastErrorMessage(
-                    fallback:
-                        "`TorLwdConn.fetchTransaction` failed with unknown error"
-                )
+                lastErrorMessage(fallback: "`TorLwdConn.fetchTransaction` failed with unknown error")
             )
         }
 
@@ -184,5 +179,82 @@ public class TorLwdConn {
                 ),
             status: isNotMined ? .notInMainChain : .mined(Int(height))
         )
+    }
+    
+    /// Gets a lightwalletd server info
+    /// - Returns: LightWalletdInfo
+    func getInfo() throws -> LightWalletdInfo {
+        let infoPtr = zcashlc_tor_lwd_conn_get_info(conn)
+        
+        guard let infoPtr else {
+            throw ZcashError.rustTorLwdGetInfo(
+                lastErrorMessage(fallback: "`TorLwdConn.getInfo` failed with unknown error")
+            )
+        }
+        
+        defer { zcashlc_free_boxed_slice(infoPtr) }
+
+        let slice = infoPtr.pointee
+        guard let rawPtr = slice.ptr else {
+            throw ZcashError.rustTorLwdGetInfo("`TorLwdConn.getInfo` Null pointer in FfiBoxedSlice")
+        }
+        
+        let buffer = UnsafeBufferPointer<UInt8>(start: rawPtr, count: Int(slice.len))
+        let data = Data(buffer: buffer)
+        
+        do {
+            let info = try LightdInfo(serializedBytes: data)
+            return info
+        } catch {
+            throw ZcashError.rustTorLwdGetInfo("`TorLwdConn.getInfo` Failed to decode protobuf LightdInfo: \(error)")
+        }
+    }
+
+    /// Gets a block at the chain tip of the blockchain
+    /// - Returns: Block
+    func latestBlock() throws -> BlockID {
+        var height: UInt32 = 0
+        
+        let blockIDPtr = zcashlc_tor_lwd_conn_latest_block(conn, &height)
+        
+        guard let blockIDPtr else {
+            throw ZcashError.rustTorLwdLatestBlockHeight(
+                lastErrorMessage(fallback: "`TorLwdConn.latestBlockHeight` failed with unknown error")
+            )
+        }
+        
+        defer { zcashlc_free_boxed_slice(blockIDPtr) }
+        
+        return BlockID(height: BlockHeight(height))
+    }
+    
+    /// Gets a tree state for a given height
+    /// - Parameter height: heght for what a tree state is requested
+    /// - Returns: TreeState
+    func getTreeState(height: BlockHeight) throws -> TreeState {
+        let treeStatePtr = zcashlc_tor_lwd_conn_get_tree_state(conn, UInt32(height))
+        
+        guard let treeStatePtr else {
+            throw ZcashError.rustTorLwdGetTreeState(
+                lastErrorMessage(fallback: "`TorLwdConn.getTreeState` failed with unknown error")
+            )
+        }
+        
+        defer { zcashlc_free_boxed_slice(treeStatePtr) }
+
+        let slice = treeStatePtr.pointee
+        guard let rawPtr = slice.ptr else {
+            throw ZcashError.rustTorLwdGetTreeState("`TorLwdConn.getTreeState` Null pointer in FfiBoxedSlice")
+        }
+        
+        let buffer = UnsafeBufferPointer<UInt8>(start: rawPtr, count: Int(slice.len))
+        let data = Data(buffer: buffer)
+        
+        do {
+            let treeState = try TreeState(serializedBytes: data)
+            return treeState
+        } catch {
+            throw ZcashError.rustTorLwdGetTreeState("`TorLwdConn.getTreeState` Failed to decode protobuf TreeState: \(error)")
+        }
     }
 }